Put browser.go in hit.go
